pyproject.toml
[tool.poetry]
name = "darkchess_poetry"
version = "0.1.0"
description = "darkchess"
authors = ["william ermlick wermlickk@gmu.edu"]
readme = "README.md"



[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-chess = "^1.999"
gym = "^0.17.2"
numpy = "^1.20.0"
ipython = "^8.18.0"
ipykernel = "^6.26.0"
stockfish = "^3.28.0"
open-spiel = "^1.4"
pandas = "^2.1.3"



[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"
ruff = "^0.0.256"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line_length = 110
target-version = ["py310"]
types = "python"

[tool.ruff]
fix = true
src = ["src"]
line-length = 110
target-version = "py310"
ignore = ["E501", "F841", "F401","E402","E712","F821", "E722"]
select = ["I","E","F"]

[tool.ruff.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 10.
max-complexity = 15

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
exclude = "src/(tests|preprocessing)"

[tool.pylint.main]
ignore = ["E1101"]

[tool.isort]
profile = "black"
line_length = 120